LPS-92465 Move into new packages
