+++ modules/apps/portal-remote/portal-remote-jaxrs-security-impl/src/main/java/com/liferay/portal/remote/jaxrs/security/internal/container/AccessControlledContainerRequestResponseFilterExceptionMapper.java
@@ -12,13 +12,13 @@
  * details.
  */
 
-package com.liferay.portal.remote.jaxrs.security.internal.access.control;
+package com.liferay.portal.remote.jaxrs.security.internal.container;
 
 import com.liferay.petra.reflect.AnnotationLocator;
 import com.liferay.portal.kernel.security.access.control.AccessControlUtil;
 import com.liferay.portal.kernel.security.access.control.AccessControlled;
 import com.liferay.portal.kernel.security.auth.AccessControlContext;
-import com.liferay.portal.remote.jaxrs.security.internal.response.ForbiddenEntity;
+import com.liferay.portal.remote.jaxrs.security.internal.entity.ForbiddenEntity;
 import com.liferay.portal.security.access.control.AccessControlAdvisor;
 import com.liferay.portal.security.access.control.AccessControlAdvisorImpl;
